build.sh

+++ build-pwa-git-web.sh
@@ -1,12 +1,22 @@
+if [ ${PWD##*/} != "cine.pwa" ];
+  then
+  echo "please run at cine.pwa."
+  exit 2
+fi
 echo "npm install start..."
 cnpm install
 echo "npm install done"
 npm run-script prod
 echo "npm prod done"
+if [ ! -d "../cine.web" ];
+  then
+  echo "not find cine.web."
+  exit 3
+fi
 cd ../cine.web
 echo "update cine.web git code start..."
 git pull origin master
-echo "update cine.web code done."
+echo "update cine.web git code done."
 rm -rf webapp
 mkdir webapp
 cp -R ../cine.pwa/build/* ./webapp
